+++ Mindbox.I18n.Templates/LocalizationTemplateParametersExtensions.cs
@@ -30,24 +30,13 @@ public static class LocalizationTemplateParametersExtensions
 			)));
 	}
 
-	public static IModelField ToModelField(this ParameterField value, ILocale locale)
+	public static LocalizationTemplateParameters WithField(
+		this LocalizationTemplateParameters localizationTemplateParameters,
+		string fieldName,
+		Func<ILocale, string> valueProvider)
 	{
-		return new ModelField(
-			value.Name,
-			value.Value.ToModelValue(locale)
-		);
+		return  localizationTemplateParameters
+			.WithField(fieldName, new PrimitiveParameter(valueProvider));
 	}
 
-	public static IModelValue ToModelValue(this ParameterValue value, ILocale locale) => value switch
-	{
-		CompositeParameter composite => composite.ToModelValue(locale),
-		PrimitiveParameter primitive => primitive.ToModelValue(locale),
-			_ => throw new InvalidCastException()
-	};
-
-	public static IModelValue ToModelValue(this PrimitiveParameter value, ILocale locale) =>
-		new PrimitiveModelValue(value.ValueProvider(locale));
-
-	public static IModelValue ToModelValue(this CompositeParameter value, ILocale locale) =>
-		new CompositeModelValue(value.Fields.Select(field => field.ToModelField(locale)));
 }
